feat: rename Recovery to AccountRecovery.
logion-network/logion-internal#1266
