+++ packages/client/src/index.ts
@@ -11,6 +11,8 @@ export {
 };
 
 export * from './assertions.js';
+export * from './AccountRecovery.js';
+export { LegalOfficerDecision, LoRecoveryClient, ProtectionRequest, ProtectionRequestStatus, UpdateParameters, UserActionParameters, CreateProtectionRequest } from './AccountRecoveryClient.js';
 export * from './AuthenticationClient.js';
 export * from './AxiosFactory.js';
 export * from './Balance.js';
@@ -31,8 +33,6 @@ export * from './NetworkState.js';
 export * from './PSP34.js';
 export * from './Polling.js';
 export * from './Public.js';
-export * from './Recovery.js';
-export { LegalOfficerDecision, LoRecoveryClient, ProtectionRequest, ProtectionRequestStatus, UpdateParameters, UserActionParameters, CreateProtectionRequest } from './RecoveryClient.js';
 export * from './SecretRecovery.js';
 export * from './SecretRecoveryClient.js';
 export * from './SharedClient.js';
